Round up score

+++ tweets-ui/src/main/scala/com/kafka/experiments/tweetsui/newsletter/NewsletterTweetDraft.scala
@@ -29,5 +29,5 @@ case class NewsletterTweet(
     url: String,
     createdAt: String,
     category: String,
-    score: Double = -1
+    score: Long = -1
 )
